+++ src/intervals/rounding.jl
@@ -258,12 +258,12 @@ This gives the narrowest possible interval.
 `prevfloat` and `nextfloat` to achieve directed rounding. This creates an interval of width 2`eps`.
 """ ->
 
-get_interval_rounding() = interval_parameters.rounding
+get_interval_rounding() = parameters.rounding
 
 function set_interval_rounding(mode)
     if mode ∉ [:wide, :narrow]
         throw(ArgumentError("Only possible interval rounding modes are `:wide` and `:narrow`"))
     end
 
-    interval_parameters.rounding = mode  # a symbol
+    parameters.rounding = mode  # a symbol
 end